+++ src/easely/cli.py
@@ -29,6 +29,7 @@ from easely import __version__, logging_
 from easely.__qt__ import bootstrap_window
 from easely.gui import ProgramBrowser, SessionDirectory, SlideShow
 from easely import face
+from easely import pdf
 from easely import tasks
 from easely.program import Program
 
@@ -163,9 +164,16 @@ class CliArgumentParser(argparse.ArgumentParser):
         rasterize.add_argument("--input-dir", type=str,
             default=tasks.RasterizeDefaults.input_dir,
             help="the input folder containing the pdf files to be rastered")
+        rasterize.add_argument("--targets", nargs="+", type=int,
+            default=tasks.RasterizeDefaults.targets,
+            help="the list of target ids to be processed")
         rasterize.add_argument("--output-dir", type=str,
             default=tasks.RasterizeDefaults.output_dir,
             help="the output folder for the generated png file(s)")
+        rasterize.add_argument("--tool", type=pdf.Raster,
+            choices=[tool.value for tool in pdf.Raster],
+            default=tasks.RasterizeDefaults.tool.value,
+            help="the rastering tool to be used for the conversion")
         rasterize.add_argument("--target-width", type=int,
             default=tasks.RasterizeDefaults.target_width,
             help="the target width for the generated png file(s)")